refactor: add GooglePaySection
